Refactor aspiration controller heuristics
